model_api/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy 